+++ packages/cadl-ranch-specs/http/parameters/body-optionality/mockapi.ts
@@ -1,29 +1,68 @@
-import { passOnSuccess, mockapi } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, MockRequest } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
+function createServerTests(uri: string, data: any) {
+  return passOnSuccess({
+    uri,
+    method: "post",
+    request: {
+      body: data,
+    },
+    response: {
+      status: 204,
+    },
+    handler: (req: MockRequest) => {
+      req.expect.bodyEquals({ name: "foo" });
+      return { status: 204 };
+    },
+    kind: "MockApiDefinition",
+  });
+}
 
-Scenarios.Parameters_BodyOptionality_requiredExplicit = passOnSuccess(
-  mockapi.post("/parameters/body-optionality/required-explicit", (req) => {
-    req.expect.bodyEquals({ name: "foo" });
-    return { status: 204 };
-  }),
+Scenarios.Parameters_BodyOptionality_requiredExplicit = createServerTests(
+  "/parameters/body-optionality/required-explicit",
+  {
+    name: "foo",
+  },
 );
 
 Scenarios.Parameters_BodyOptionality_OptionalExplicit = passOnSuccess([
-  mockapi.post("/parameters/body-optionality/optional-explicit/set", (req) => {
-    req.expect.bodyEquals({ name: "foo" });
-    return { status: 204 };
-  }),
-  mockapi.post("/parameters/body-optionality/optional-explicit/omit", (req) => {
-    req.expect.rawBodyEquals(undefined);
-    return { status: 204 };
-  }),
+  {
+    uri: "/parameters/body-optionality/optional-explicit/set",
+    method: "post",
+    request: {
+      body: {
+        name: "foo",
+      },
+    },
+    response: {
+      status: 204,
+    },
+    handler: (req: MockRequest) => {
+      req.expect.bodyEquals({ name: "foo" });
+      return { status: 204 };
+    },
+    kind: "MockApiDefinition",
+  },
+  {
+    uri: "/parameters/body-optionality/optional-explicit/omit",
+    method: "post",
+    request: {},
+    response: {
+      status: 204,
+    },
+    handler: (req: MockRequest) => {
+      req.expect.rawBodyEquals(undefined);
+      return { status: 204 };
+    },
+    kind: "MockApiDefinition",
+  },
 ]);
 
-Scenarios.Parameters_BodyOptionality_requiredImplicit = passOnSuccess(
-  mockapi.post("/parameters/body-optionality/required-implicit", (req) => {
-    req.expect.bodyEquals({ name: "foo" });
-    return { status: 204 };
-  }),
+Scenarios.Parameters_BodyOptionality_requiredImplicit = createServerTests(
+  "/parameters/body-optionality/required-implicit",
+  {
+    name: "foo",
+  },
 );